.history/requirements_20231126185424.txt
absl-py==2.0.0
addict==2.4.0
albumentations==1.3.1
APScheduler==3.10.4
basicsr==1.4.2
black==23.9.1
cachetools==5.3.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.0
Cython==3.0.3
easydict==1.10
facexlib==0.3.0
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.43.0
future==0.18.3
gfpgan==1.3.8
google-auth==2.23.2
google-auth-oauthlib==1.1.0
grpcio==1.59.0
humanfriendly==10.0
imageio==2.31.5
importlib-metadata==6.8.0
insightface==0.7.3
isort==5.12.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.0
lmdb==1.4.1
Markdown==3.4.4
matplotlib==3.8.0
minio==7.1.17
mypy-extensions==1.0.0
numba==0.58.0
oauthlib==3.2.2
onnx==1.14.0
onnxruntime-gpu==1.15.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
prettytable==3.9.0
protobuf==4.23.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
tb-nightly==2.15.0a20231005
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tomli==2.0.1
tqdm==4.66.1
tzdata==2023.3
tzlocal==5.1
wcwidth==0.2.8
Werkzeug==3.0.0
yapf==0.40.2
zipp==3.17.0
moviepy==1.0.3